Dump of port draft to Qt6/KF6 (missing Structures tool for now)
* contains the complete dump of the “it builds, starts and does not crash
  on simple usage” changes.
* continuously rebased to master branch, the latest current Qt5/KF5-based
* Structures tool disabled from build, needs QtScript port

This commit will at the time of the real Qt6/KF6 port be properly split
into the different aspects of porting.
For now it serves to hold the door open while still on the Qt6/KF6 side.

+++ kasten/gui/liboktetawidgets/bytearrayutf8stringvalidator.cpp
@@ -56,7 +56,7 @@ QByteArray ByteArrayUtf8StringValidator::toByteArray(const QString& string) cons
 
 QString ByteArrayUtf8StringValidator::toString(const QByteArray& byteArray) const
 {
-    const int byteArraySize = std::min(byteArray.size(), maxLength());
+    const int byteArraySize = std::min(static_cast<int>(byteArray.size()), maxLength());
 
     QString result = m_encoder.encodeAsString(byteArray.constData(), byteArraySize, m_decoder.textCodec());
